% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.2dmixplot}
\alias{QtAC.2dmixplot}
\title{2dmixplot}
\usage{
QtAC.2dmixplot(
  Mat,
  prop1,
  prop2,
  Save = FALSE,
  filename = paste("2dmixplot", prop1, prop2, sep = "_")
)
}
\arguments{
\item{Mat}{data frame containing the systemic variables}

\item{prop1}{variable on x-axis ("potential","connectedness","resilience")}

\item{prop2}{variable on y-axis ("potential","connectedness","resilience")}

\item{Save}{If Save = TRUE, the 2D plot will be saved in a PNG file.}

\item{filename}{If Save = TRUE, the network will be saved in a file called filename.}
}
\value{
2D plot and, if Save = TRUE, a PNG file containing the plot
}
\description{
This function plots two selected systemic variables w.r.t. each other.
}
