% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.2dplot}
\alias{QtAC.2dplot}
\title{2dplot}
\usage{
QtAC.2dplot(
  Mat,
  prop = NULL,
  time_int = NULL,
  Save = FALSE,
  filename = "2dplot"
)
}
\arguments{
\item{Mat}{dataframe containing the systemic variables}

\item{prop}{If prop = NULL, the three systemic variables are plotted w.r.t time in one plots each. If prop = "potential", "connectedness", or "resilience", only the selected systemic variable is plotted w.r.t time.}

\item{time_int}{vector containing start time, end time, and step size to define the xaxis}

\item{Save}{If Save = TRUE, the 2D plot will be saved in a PNG file.}

\item{filename}{If Save = TRUE, the network will be saved in a file called filename.}
}
\value{
2D plot and, if Save = TRUE, a PNG file containing the plots.
}
\description{
This function plots graphs of potential, connectedness and resilience with respect to time.
}
