% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.3dplot}
\alias{QtAC.3dplot}
\title{3dplot}
\usage{
QtAC.3dplot(Mat, Mat_points = FALSE)
}
\arguments{
\item{Mat}{data frame containing the three systemic variables}

\item{Mat_points}{If Mat_points = TRUE, the maturation points are visible.}
}
\value{
3D plot
}
\description{
3D plot of the three systemic variables w.r.t each other.
}
