% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.maturation}
\alias{QtAC.maturation}
\title{maturation}
\usage{
QtAC.maturation(result_mtx, res_norm = "continous", res_stand = "maxweight")
}
\arguments{
\item{result_mtx}{list of adjacency matrices and significance matrices}

\item{res_norm}{normalization variant of the Laplacian matrices ("continous", "symmetric"). If res_norm = "symmetric", \eqn{L_{out} = c \cdot D^{-\frac{1}{2}}_{out} \left(D_{out}- A \right)} and \eqn{L_{in} = c \cdot \left(D_{in} - A \right) D^{-\frac{1}{2}}_{in}}. If res_norm = "symmetric", \eqn{L_{out} = c \cdot D^{-\frac{1}{2}}_{out} (D_{out}-A) D^{-\frac{1}{2}}_{out}} and \eqn{L_{in} =  c\cdot D^{-\frac{1}{2}}_{in} (D_{in}-A)D^{-\frac{1}{2}}_{in}}.}

\item{res_stand}{standardization constant c of the Laplacian matrices ("none", "maxweight", "maxweight2", "nodes", "maxweightnodes"). Let \eqn{N} be the number of nodes of the underlying graph and \eqn{M} its maximal edge weight. If res_stand = "none", \eqn{c = 1}. If res_stand = "maxweight", \eqn{c = \frac{1}{\sqrt{M}}}. If res_stand = "maxweight2", \eqn{c = \frac{1}{M}}. If res_stand = "nodes", \eqn{c = \frac{\sqrt{N-1}}{N}}. If res_stand = "maxweightnodes", \eqn{c = \frac{\sqrt{N-1}}{N \cdot \sqrt{M}}}.}
}
\value{
dataframe containing the three systemic variables (potential, connectedness, and resilience) of each adjacency matrix
}
\description{
This function computes the three systemic variables (potential, connectedness, and resilience) of each adjacency matrix.
}
