% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC}
\alias{QtAC}
\title{QtAC (Main function)}
\usage{
QtAC(
  Data,
  num_timepoints = 5,
  JavaPath,
  num_PermCheck = 1000L,
  k = 1L,
  k_tau = 1L,
  l = 1L,
  l_tau = 1L,
  delay = 1L
)
}
\arguments{
\item{Data}{data array containing time series of the system's components' abundance data}

\item{num_timepoints}{length of the time windows of abundance data serving as basis of the transfer entropy calculations}

\item{JavaPath}{path of the file "infodynamics.jar"}

\item{num_PermCheck}{number of surrogate samples to bootstrap to generate the distribution in the significance test}

\item{k}{embedding length of destination past history to consider}

\item{k_tau}{embedding delay for the destination variable}

\item{l}{embedding length of source past history to consider}

\item{l_tau}{embedding delay for the source variable}

\item{delay}{time lag between last element of source and destination next value}
}
\value{
list of two lists (adjacency and corresponding significance matrices)
}
\description{
This function calculates the transfer entropy between two species each for shifting time windows of fixed length.
The output is a list of adjacency matrices and the corresponding significance matrices.
}
