% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.signfactor_CTE}
\alias{QtAC.signfactor_CTE}
\title{.signfactor_CTE}
\usage{
QtAC.signfactor_CTE(result_mtx, signfac = 0.05)
}
\arguments{
\item{result_mtx}{list of matrices containing summands of collective transfer entropies (one row per species) and corresponding significance matrices}

\item{signfac}{significance level}
}
\value{
list of two lists (vectors of significant collective transfer entropies and significance matrices)
}
\description{
This function sets all transfer entropies to 0 whose p-value is above the predefined significance level.
}
