% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.TXT.reader}
\alias{QtAC.TXT.reader}
\title{TXT-reader}
\usage{
QtAC.TXT.reader(filename, col_names = FALSE, row_names = FALSE)
}
\arguments{
\item{filename}{path of the file you want to import}

\item{col_names}{logical operator. TRUE if the file contains column names, FALSE if it does not}

\item{row_names}{logical operator. TRUE if the file contains row names, FALSE if it does not}
}
\value{
data array. If no column or row names were given, column or row names of the form "t_ " and "C_ " respectively are added.
}
\description{
This function is used to import the data in R. The data should be in a tab-separated file with or without column/row names. Columns should contain time points, rows the system's components.
}
