% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.WriteAdjSgn}
\alias{QtAC.WriteAdjSgn}
\title{WriteAdjSgn}
\usage{
QtAC.WriteAdjSgn(
  result_mtx,
  num_mtx,
  file_name_adj = "adjacency_matrix",
  file_name_sgn = "significance_matrix"
)
}
\arguments{
\item{result_mtx}{list of adjacency and significance matrices}

\item{num_mtx}{number of the adjacency and significance matrix you want to write to a CSV file}

\item{file_name_adj}{name of the file to which the adjacency matrix is written}

\item{file_name_sgn}{name of the file to which the significance matrix is written}
}
\value{
two CSV files, one containing the selected adjacency, one the significance matrix
}
\description{
This function writes a selected adjacency and significance matrix to CSV files.
}
