% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC_AIS}
\alias{QtAC_AIS}
\title{QtAC_AIS}
\usage{
QtAC_AIS(
  data,
  num_timepoints,
  javapath,
  noise_level = "1e-20",
  num_permcheck = 1000L,
  k = 1L,
  k_tau = 1L,
  mode = c("average", "local"),
  save = FALSE,
  filename = "result_QtAC"
)
}
\arguments{
\item{data}{data array containing time series of the system's components' abundance data}

\item{num_timepoints}{length of the time windows of abundance data serving as basis of the active information storage estimations}

\item{javapath}{path of the file "MTinfodynamics.jar"}

\item{noise_level}{amount of random Gaussian noise added in the estimation}

\item{num_permcheck}{number of surrogate samples to bootstrap to generate the distribution in the significance test}

\item{k}{embedding length of destination past history to consider}

\item{k_tau}{embedding delay}

\item{mode}{Active information storage is either averaged over time windows ("average") or local values are estimated ("local"). In the latter case, the value of num_timepoints is ignored.}

\item{save}{If save=TRUE, the output is saved in a file called filename.}
}
\value{
list of two vectors (AIS of all species and corresponding significance values)
}
\description{
This function calculates the active information storage of each species for shifting time windows of fixed length or local values.
The output is a list of adjacency matrices and the corresponding significance matrices.
}
