% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.2dmixplot}
\alias{QtAC.2dmixplot}
\title{2dmixplot}
\usage{
QtAC.2dmixplot(
  mat,
  prop1,
  prop2,
  save = FALSE,
  filename = paste("2dmixplot_", prop1, "_", prop2)
)
}
\arguments{
\item{prop1}{variable on x-axis ("potential","connectedness","resilience")}

\item{prop2}{variable on y-axis ("potential","connectedness","resilience")}

\item{save}{If Save = TRUE, the 2D plot will be saved in a PNG file.}

\item{filename}{If Save = TRUE, the network will be saved in a file called filename.}

\item{Mat}{data frame containing a time series of systemic variables}
}
\value{
2D plot and, if Save = TRUE, a PNG file containing the plot
}
\description{
This function plots two selected systemic variables w.r.t. each other. A curve is interpolated via a piecewise cubic spline.
}
