% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.2dplot}
\alias{QtAC.2dplot}
\title{2dplot}
\usage{
QtAC.2dplot(
  mat,
  prop = NULL,
  time_int = NULL,
  time = "time",
  save = FALSE,
  filename = "2dplot"
)
}
\arguments{
\item{mat}{dataframe containing a time series of systemic variables}

\item{prop}{If prop = NULL, the three systemic variables are plotted w.r.t time in one plot each. If prop = "potential", "connectedness", or "resilience", only the selected systemic variable is plotted w.r.t time.}

\item{time_int}{vector containing start time, end time, and step size to define the xaxis}

\item{time}{unit (i.e. "years", "steps",...)}

\item{save}{If save = TRUE, the 2D plot will be saved in a PNG file.}

\item{filename}{If save = TRUE, the network will be saved in a file called filename.}
}
\value{
2D plot and, if save = TRUE, a PNG file containing the plots.
}
\description{
This function plots potential, connectedness and resilience with respect to time. Curves are interpolated via a piecewise cubic spline.
}
