% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.3dplot}
\alias{QtAC.3dplot}
\title{3dplot}
\usage{
QtAC.3dplot(mat, mat_points = FALSE)
}
\arguments{
\item{mat_points}{If mat_points = TRUE, the maturation points are visible.}

\item{Mat}{data frame containing a time series of the three systemic variables}
}
\value{
3D plot
}
\description{
3D plot of the three systemic variables w.r.t each other. A curve is interpolated via a piecewise cubic spline.
}
