% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.maturation}
\alias{QtAC.maturation}
\title{maturation}
\usage{
QtAC.maturation(result_mtx, res_stand = "maxweight")
}
\arguments{
\item{result_mtx}{list of adjacency matrices and significance matrices}

\item{res_stand}{standardization constant c of the Laplacian matrices ("none", "maxweight", "nodes", "maxweightnodes"). Let \eqn{N} be the number of nodes of the underlying graph and \eqn{M} its maximal edge weight. If res_stand = "none", \eqn{c = 1}. If res_stand = "maxweight", \eqn{c = \frac{1}{\sqrt{M}}}. If res_stand = "nodes", \eqn{c = \frac{\sqrt{N-1}}{N}}. If res_stand = "maxweightnodes", \eqn{c = \frac{\sqrt{N-1}}{N \cdot \sqrt{M}}}.}
}
\value{
dataframe containing the three systemic variables (potential, connectedness, and resilience) of each adjacency matrix
}
\description{
This function computes the three systemic variables (potential, connectedness, and resilience) of each adjacency matrix.
}
