% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.network}
\alias{QtAC.network}
\title{network}
\usage{
QtAC.network(
  result_mtx,
  num_mtx = 1,
  edge_label = FALSE,
  dec = 2,
  layout = "nicely",
  edge_width = 3,
  arrow_width = 5,
  col_node = "palegreen3",
  col_edge = "steelblue3",
  vertex_label = "short",
  title = paste("Network of Adjacency matrix ", times[num_mtx]),
  save = FALSE,
  filename = paste("network_", num_mtx, sep = "")
)
}
\arguments{
\item{result_mtx}{list of adjacency matrices and significance matrices}

\item{num_mtx}{number of the adjacency matrix you want to plot (default = 1)}

\item{edge_label}{If edge_label = TRUE, the weight of the edges are plotted next to the edges.}

\item{dec}{number of decimal digits in the edge labels}

\item{layout}{layout format ("circle","star","fruchterman.reingold","grid","nicely")}

\item{edge_width}{muliplicator for the width of the edges}

\item{arrow_width}{muliplicator for the width of the arrows}

\item{col_node}{color of the vertices}

\item{col_edge}{color of the edges}

\item{vertex_label}{If vertex_label = "short", the first 4 letters of the components' names will be used as vertex labels, if vertex_label = "long", the whole names will be used as vertex labels. Via vertex_label =  c(...), customized names can be used as vertex labels.}

\item{title}{title of the network}

\item{save}{If save = TRUE, the network will be saved as a PNG file.}

\item{filename}{If save = TRUE, the network will be saved in a file called filename.}
}
\value{
network plot and, if save = TRUE, a PNG file containing the plot
}
\description{
This function plots a selected adjacency matrix as network.
}
