% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC}
\alias{QtAC}
\title{QtAC (Main function)}
\usage{
QtAC(
  data,
  num_timepoints,
  javapath,
  noise_level = "1e-20",
  num_permcheck = 1000L,
  k = 1L,
  k_tau = 1L,
  l = 1L,
  l_tau = 1L,
  delay = 1L,
  save = FALSE,
  filename = "result_QtAC"
)
}
\arguments{
\item{data}{data array containing time series of the system's components' abundance data}

\item{num_timepoints}{length of the time windows of abundance data serving as basis of the transfer entropy estimations}

\item{javapath}{path of the file "MTinfodynamics.jar"}

\item{noise_level}{amount of random Gaussian noise added in the estimation}

\item{num_permcheck}{number of surrogate samples to bootstrap to generate the distribution in the significance test}

\item{k}{embedding length of destination past history to consider}

\item{k_tau}{embedding delay for the destination variable}

\item{l}{embedding length of source past history to consider}

\item{l_tau}{embedding delay for the source variable}

\item{delay}{time lag between last element of source and destination next value}

\item{save}{If save=TRUE, the output is saved in a file called filename.}
}
\value{
list of two lists (adjacency and corresponding significance matrices)
}
\description{
This function calculates the transfer entropy between two species each for shifting time windows of fixed length.
The output is a list of adjacency matrices and the corresponding significance matrices.
}
