% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtAC_package.R
\name{QtAC.signfactor}
\alias{QtAC.signfactor}
\title{signfactor}
\usage{
QtAC.signfactor(result_mtx, signfac = 0.05)
}
\arguments{
\item{result_mtx}{list of adjacency and significance matrices}

\item{signfac}{significance level}
}
\value{
list of two lists (adjacency matrix containing only significant transfers and significance matrices)
}
\description{
This function sets all entries in the adjacency matrices to 0 whose p-value is above the predefined significance level.
}
